pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llmsurvey"
version = "0.1.0"
description = "Measure how well LLMs simulate human survey respondents"
requires-python = ">=3.11"
dependencies = [
  "replicate>=0.34",
  "click>=8.1",
  "pydantic>=2.0",
  "jinja2>=3.1",
  "pyyaml>=6.0",
  "requests>=2.31",
  "scipy>=1.12",
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "python-dotenv>=1.0",
]

[project.scripts]
llmsurvey = "llmsurvey.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["llmsurvey"]
